Refactor: Enhance quantum processing and data ingestion for improved pattern analysis
This commit enhances the quantum processing capabilities and data ingestion methods within the SEP engine, focusing on improved pattern analysis and metric calculation.

- **Quantum Processing Enhancements:**
  - Added `entropy` field to `QFHResult` struct in `src/quantum/qfh.h` to capture information entropy of bit patterns.
  - Modified `src/quantum/quantum_processor_qfh_common.cpp` to calculate and utilize entropy in pattern analysis, improving coherence calculations. Normalizes input patterns and refines entropy calculation using Shannon's formula, clamping the result between 0 and 1.
  - Linked `CUDA::cufft` library in `src/quantum/CMakeLists.txt` to enable cuFFT support for potential future GPU-accelerated quantum computations.
  - Corrected double-promotion issue in `src/core/dag_graph_quant.cpp` by casting `node.generation` to float before logarithmic calculation.

- **Data Ingestion Improvements:**
  - Implemented `ingestFile` method in `src/core/engine.h` and `src/core/engine.cpp` to allow ingestion of data from files, leveraging the `DataParser` for pattern extraction.
  - Enhanced `DataParser` in `src/core/data_parser.cpp` to improve data format detection and parsing, including handling of partial JSON data and deriving coherence from variance in binary data.
  - Modified `PatternMetricEngine` in `src/quantum/pattern_metric_engine.cpp` and `src/quantum/pattern_metric_engine.h` to use a mutex-protected stream buffer for processing data in chunks, improving thread safety and streaming capabilities. The `processBuffer` method now extracts patterns directly from the stream buffer and calculates quantum states.

- **Compatibility and Code Quality:**
  - Fixed CUDA macro definitions in `src/compat/cuda_unified_fix.h` to avoid naming conflicts.
  - Addressed unused return value warnings in `src/core/engine.cpp` by casting the result of `fflush` to `void`.
  - Removed unused variable `cleanup_threshold` in `src/memory/memory_tier.cpp`.

+++ src/core/dag_graph_quant.cpp
@@ -103,7 +103,7 @@ void DagGraph::calculateAlpha() {
         float baseAlpha = node.coherence;
         
         // Generation bonus (more evolved patterns have higher alpha)
-        float generationBonus = std::log1p(node.generation) * 0.1f;
+        float generationBonus = std::log1p(static_cast<float>(node.generation)) * 0.1f;
         
         // Stability factor
         float stabilityFactor = node.stability;
